+++ src/vl_mamba/evaluation/evaluate_textvqa.py
@@ -291,7 +291,7 @@ def pythia_generate(
         # num_beams=5,
     )
 
-    with torch.autocast("cuda", dtype=torch.bfloat16):
+    with torch.autocast("cuda", dtype = torch.float16):  # dtype=torch.bfloat16
         outputs = model.generate(
             input_ids=input_ids,
             pixel_values=[pixel_values],
